backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
azure-ai-documentintelligence==1.0.0b1
azure-common==1.1.28
azure-core==1.36.0
azure-cosmos==4.5.1
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-servicebus==7.11.4
azure-storage-blob==12.19.0
beautifulsoup4==4.12.2
black==24.1.1
blis==0.7.11
cachetools==6.2.1
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cloudpathlib==0.16.0
confection==0.1.5
coverage==7.10.7
cryptography==46.0.3
cymem==2.0.11
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.109.0
flake8==7.0.0
google-api-core==2.28.1
google-auth==2.42.1
googleapis-common-protos==1.71.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.26.0
idna==3.11
iniconfig==2.1.0
isodate==0.7.2
Jinja2==3.1.6
joblib==1.5.2
langcodes==3.5.0
language_data==1.3.0
lxml==6.0.2
marisa-trie==1.3.1
MarkupSafe==3.0.3
mccabe==0.7.0
msal==1.34.0
msal-extensions==1.3.1
murmurhash==1.0.13
mypy==1.8.0
mypy_extensions==1.1.0
nltk==3.8.1
numpy==1.26.4
openai==1.10.0
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
preshed==3.0.10
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pycparser==2.23
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pyflakes==3.2.0
PyJWT==2.10.1
pypdf==4.0.1
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
python-docx==1.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.3
regex==2025.11.3
requests==2.31.0
aiohttp==3.9.1
rsa==4.9.1
sentence-transformers==2.3.1
-e git+https://AIAgentCapabilityPOC@dev.azure.com/AIAgentCapabilityPOC/Agent%20POCs/_git/Agent%20POCs@584bb090199fe45d331abb741104fac7a5e4fba5#egg=script_to_doc&subdirectory=backend
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.8
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.35.1
thinc==8.2.5
tomli==2.3.0
tqdm==4.67.1
typer==0.9.4
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.27.0
uvloop==0.22.1
wasabi==1.1.3
watchfiles==1.1.1
weasel==0.3.4
websockets==15.0.1

# Document conversion (User Story 117: File Conversion Service)
# Note: LibreOffice must be installed on the system for conversions to work
# Install: apt-get install -y libreoffice-writer libreoffice-impress libreoffice-core
